requirements.txt
flask
gunicorn
tensorflow-cpu
keras
numpy
pandas
opencv-python-headless
scikit-learn
pillow